requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
async-timeout==4.0.2
attrs==19.3.0
Automat==0.8.0
awscli==1.18.69
blinker==1.4
botocore==1.16.19
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.1.0
Click==7.0
cloud-init==22.1
cmake==3.26.3
colorama==0.4.3
coloredlogs==15.0.1
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==2.8
cycler==0.11.0
dbus-python==1.2.16
diffusers==0.16.1
distro==1.4.0
distro-info===0.23ubuntu1
docutils==0.16
entrypoints==0.3
exceptiongroup==1.1.1
fastapi==0.95.2
ffmpy==0.3.0
filelock==3.12.0
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gradio==3.33.1
gradio-client==0.2.5
h11==0.14.0
httpcore==0.17.2
httplib2==0.14.0
httpx==0.24.1
huggingface-hub==0.15.1
humanfriendly==10.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==5.12.0
incremental==16.10.1
Jinja2==2.10.1
jmespath==0.9.4
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
linkify-it-py==2.0.2
lit==16.0.5
markdown-it-py==2.2.0
MarkupSafe==1.1.0
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
more-itertools==4.2.0
mpmath==1.3.0
multidict==6.0.4
netifaces==0.10.4
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
olefile==0.46
onnx==1.14.0
onnxruntime==1.15.0
opencv-python==4.7.0.72
orjson==3.9.0
packaging==23.1
pandas==2.0.2
pexpect==4.6.0
Pillow==7.0.0
protobuf==4.23.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycocotools==2.0.6
pydantic==1.10.8
pydub==0.25.1
Pygments==2.15.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-multipart==0.0.6
pytz==2023.3
PyYAML==5.3.1
regex==2023.5.5
requests==2.22.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.0
s3transfer==0.3.3
SecretStorage==2.3.1
semantic-version==2.10.0
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sos==4.3
ssh-import-id==5.10
starlette==0.27.0
sympy==1.12
systemd-python==234
toolz==0.12.0
torch==2.0.1
tqdm==4.65.0
triton==2.0.0
Twisted==18.9.0
typing-extensions==4.6.3
tzdata==2023.3
ubuntu-advantage-tools==27.6
uc-micro-py==1.0.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.22.0
wadllib==1.3.3
websockets==11.0.3
yarl==1.9.2
zipp==3.15.0
zope.interface==4.7.1
